splitt GdprAgreement Card into smaler Components.
