+++ src/components/gdpr/GdprAgreementCard.vue
@@ -1,99 +1,30 @@
 <template>
   <v-card>
-    <v-toolbar flat>
-      <v-toolbar-title>
-        {{ $t("app.privacyagreement") }}
-      </v-toolbar-title>
-
-      <v-spacer></v-spacer>
-
-      <LanguageSwitcher />
-
-      <v-btn text @click="logout">
-        {{ $t("actions.logout") }}
-      </v-btn>
-    </v-toolbar>
-
-    <v-card-text class="pb-0">
-      <i18n path="privacyagreement.text" tag="p">
-        <template #privacyAgreement>
-          <router-link :to="{ name: 'privacy' }">{{
-            $t("app.privacyagreement")
-          }}</router-link>
-        </template>
-      </i18n>
-
-      <v-checkbox v-model="value" :label="$t('privacyagreement.checkbox')">
-      </v-checkbox>
-    </v-card-text>
-    <v-card-text> {{ $t("privacyagreement.revokeInfo") }} </v-card-text>
-
-    <v-card-actions>
-      <v-btn text color="primary" :disabled="!value || loading" @click="save">
-        {{ $t("actions.complete") }}
-      </v-btn>
-    </v-card-actions>
-
+    <CardToolbar
+      :title="$t('app.privacyagreement')"
+      :logout-action="true"
+    ></CardToolbar>
+    <GdprCardText @checkbox-updated="updateValue"></GdprCardText>
+    <GdprCardActions :value="value"></GdprCardActions>
     <FeedbackMenu />
   </v-card>
 </template>
 
 <script>
-import { log } from "@/utils/log";
-import { mdiDelete, mdiFileDocument, mdiClose, mdiRecord } from "@mdi/js";
-
 import FeedbackMenu from "@/components/FeedbackMenu";
-
-import AuthService from "@/services/auth";
-import LanguageSwitcher from "@/components/LanguageSwitcher";
+import CardToolbar from "@/components/cards/CardToolbar";
+import GdprCardText from "@/components/gdpr/GdprCardText";
+import GdprCardActions from "@/components/gdpr/GdprCardActions";
 
 export default {
   name: "GdprAgreementCard",
-  components: { FeedbackMenu, LanguageSwitcher },
+  components: { FeedbackMenu, GdprCardActions, GdprCardText, CardToolbar },
   data: () => ({
-    value: false,
-    confirmDialog: false,
-    contractFormValid: false,
-    icons: {
-      mdiFileDocument,
-      mdiDelete,
-      mdiClose,
-      mdiRecord
-    },
-    loading: false
+    value: false
   }),
   methods: {
-    logout() {
-      this.$store.dispatch("auth/LOGOUT");
-    },
-    async save() {
-      this.loading = true;
-      try {
-        await AuthService.updateSettings({
-          dsgvo_accepted: true
-        });
-
-        await this.$store.dispatch("GET_USER");
-
-        this.$router
-          .push({
-            name: "dashboard"
-          })
-          .catch(() => {});
-      } catch (error) {
-        await this.$store.dispatch("snackbar/setSnack", {
-          snack: this.$t("snackbar.error"),
-          timeout: 4000,
-          color: "warning"
-        });
-
-        // TODO: Set error state
-        log(error);
-      } finally {
-        setTimeout(() => {
-          this.loading = false;
-        }, 500);
-      }
+    updateValue(e) {
+      this.value = e;
     }
   }
 };
